Adrien De Croy's version of Wixel-SDK with dexbridge upp
This is a version of the pololu wixel SDK, modified by Adrien de Croy.
It contains a wixel app called dexbridge, and this SDK must be used to
build it.
Dexbridge, built on Adrien de Croy's dexterity-wixel, can receive and
output as text the packets from a Dexcom G4 Platinum CGM Transmitter.
It has been modified to remove the radio configuration commands (Adrien
has worked all that out in the default settings) and adds packet
filtering on Dexcom Transmitter ID.