Minor Updates
Changed my TXID in dex-drip.c for some testing.

Updated dexbridge2 to correctly set BLE ID of HM-1X.  Includes all the
sleep code, but not being activated as yet.

Created usb_serial2 to enable me to raise the HM-1X SYS/BRK pin to make
it disconnect and return to command mode for testing sleep
functionality.  Probably of no use to anyone.

usb_serial not actually changed, just saved.  It can be used to update
the firmware of the HM-1X modules.