Significant reliability changes.
Modified code to work around the apparent habit of the SDCC code NOT
being started just when the wixel was put to sleep.  Now doing a lot of
the status setting/clearing in ISR for sleep timer.
Configured timeout on loop waiting for do_sleep, so it doesn't get stuck
ad infinitum.
Tested.  All appears to be working so far.
Still some craziness when USB is disconnected after writing the wixel
file.  Not sure what it is yet.
Will probably do a full rewrite with state machine coding.  Will likely
be more reliable than this, and more efficient.