Minor adjustments.  Fixed resend of data packet
This commit fixes the continual resending of data packets and limits the
sending of data packets to 2 minutes from a Dexcom packet capture.  This
should be sufficient to deal with slow BLE reconnection after waking,
which sometimes occurs, and makes certain that we never get a negative
sleep time (now-pkt_time > sleep time).  This situation was sometimes
causing the wixel to sleep indefinitely.