Change to Beacon code
Enables waitForPacket to quit if it has sat on the first channel for
more than 3 minutes and 20 seconds.  This then sends a Beacon packet,
before it is invoked by main and getPacket again.
Ensures a beacon packet is sent even if a packet has not been received,
as in when a persons transmitter dies and they get a new one.  The new
TXID will be set when the beacon packet is received by xDrip.